requirements.txt
Django==2.2
django-redis==5.0.0
django-shortuuidfield==0.1.3
importlib-metadata==4.6.0
Markdown==3.3.4
mysqlclient==2.0.3
pytz==2021.1
redis==3.5.3
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
zipp==3.5.0
